moved Encryption class to dBHandler.js to keep it private
